Restructure api folder so that global variables available through __init__, renamed log_models to models, and app.py now queries through db to fetch data to make request. In current form, still done upon a request made, but want to separate that in a later commit.
